refactor(organon): extract energeia::metrics submodule (#2642)
Move metron tool definition and MetronExecutor to metrics.rs. Clean up
mod.rs to contain only module declarations, re-exports, registration,
and tests.
